example/python/requirements.txt
 grpcio-tools
 tzlocal
 pywin32
 tqdm